requirements.txt
regex
pillow